refactor(api): move getOrganizationDetails usecase to src/team
