+++ api/tests/organizational-entities/unit/domain/usecases/get-organization-details.usecase.test.js
@@ -1,8 +1,8 @@
-import { Organization } from '../../../../lib/domain/models/index.js';
-import { getOrganizationDetails } from '../../../../lib/domain/usecases/organizations-administration/get-organization-details.js';
-import { domainBuilder, expect, sinon } from '../../../test-helper.js';
+import { Organization } from '../../../../../src/organizational-entities/domain/models/Organization.js';
+import { getOrganizationDetails } from '../../../../../src/organizational-entities/domain/usecases/get-organization-details.usecase.js';
+import { domainBuilder, expect, sinon } from '../../../../test-helper.js';
 
-describe('Unit | UseCase | get-organization-details', function () {
+describe('Unit | Organizational Entities | Domain | UseCase | get-organization-details', function () {
   it('should return the Organization matching the given organization ID', async function () {
     // given
     const organizationId = 1234;
